# Resuse writes, initalize queues