@benaadams 2015-11-13
I didn't feel it was worthwhile to go round the Travis repeat loop for to try to get to pass; will rebase.