@benaadams 2015-11-13
Rebased and added SocketOutput