@benaadams 2015-11-13
Allocation reductions for 8k requests

Before 32,023 objects; 2,032,968 bytes

![before](http://aoa.blob.core.windows.net/aspnet/preallocate-before.png)

After 12,059 objects; 947,168 bytes

![after](http://aoa.blob.core.windows.net/aspnet/preallocate-after.png)

37% of the objects than before; 46% of the bytes of before