@benaadams 2015-11-14
Removed the allocs from `DoWriteIfNeeded` so down by another 212800