@benaadams 2015-11-14
Boom! Use `MemoryPool2` rather than `new byte[buffer.Count]` in write and `SocketOutput` alloc free