@DamianEdwards 2015-11-18
Rebased on 7e8a405 and measured. Basically the same. I really need to profile it to see if the size we're setting them to is high enough.

```
dev (7e8a405) - 499k
363 (2a32d33) - 501k (+0.4%)
```