@AspNetSmurfLab 2015-11-18
Still spending some time in resizing it seems:
![image](https://cloud.githubusercontent.com/assets/13443546/11232706/cbf53aa2-8d6a-11e5-9b15-d1fbdfc370b4.png)