@mattwarren 2015-11-18
@benaadams I'm pretty sure that screenshot is from [JetBrains DotTrace](https://www.google.co.uk/search?q=jetbrains+dottrace+screenshots+3.5&espv=2&biw=1257&bih=929&source=lnms&tbm=isch&sa=X&ved=0CAcQ_AUoAmoVChMIscS6qfWZyQIVxAcsCh0GdAVY&dpr=1#tbm=isch&q=jetbrains+dottrace+screenshot&imgrc=6596Vzm1wXxoAM%3A)