@khellang 2015-11-18
Yep, looks like and older version of dotTrace :smile: