@benaadams 2015-11-18
With 45% GC load though all bets are off on algorithmic improvements; need to get that overhead down first.