@mattwarren 2015-11-18
Out of interest, which version of the GC were these tests run under? Server, Client, Background Server, etc?