@AspNetSmurfLab 2015-11-20
Good results, nearly 10% improvement:
```
Running 10s test @ http://10.0.0.100:5001/plaintext
  32 threads and 256 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.17ms    2.70ms 181.45ms   74.48%
    Req/Sec    17.22k     1.20k   29.55k    85.90%
  5518027 requests in 10.10s, 694.64MB read
Requests/sec: 546339.52
Transfer/sec:     68.78MB
```