@davidfowl 2015-11-21
This change isn't just queue sizes it's pooling he write context and it will conflict with the socket output changes