@benaadams 2015-11-21
Yes this needs revisiting post socket output