@benaadams 2015-11-24
Update!

Rebased; renamed, changed to work with the SocketOutput changes. Results as follows....

Before WriteAsync bytes used 352,000 bytes

![write-async-before](http://aoa.blob.core.windows.net/aspnet/writeasync-higher.png)

After WriteAsync bytes used 888 bytes; 888 bytes? Crazy town!

![write-async-after](http://aoa.blob.core.windows.net/aspnet/writeasync-lowlowlow.png)

Before DoWriteIfNeeded 768,000 bytes

![DoWrite](http://aoa.blob.core.windows.net/aspnet/dowrite-higher.png)

After DoWriteIfNeeded doesn't use any bytes at all, so I have nothing to screenshot O_o

So instead, the full SocketOutput stack:

Before

![socket-output-before](http://aoa.blob.core.windows.net/aspnet/socketoutput-total-higher.png)

After

![socket-output-after](http://aoa.blob.core.windows.net/aspnet/socketoutput-total-lowlowlow.png)

The mind boggles... :dancer: