@benaadams 2015-11-25
Bug fix in this PR moving back to `ScheduleWrite` from `WriteAllPending`; however as that may cause a regression have added https://github.com/aspnet/KestrelHttpServer/pull/427 to resolve the regression.