@benaadams 2015-12-04
Still one of the big players after recent changes; nice to see `byte[]` has dropped of the list though

![write-context](http://aoa.blob.core.windows.net/aspnet/write-context.png)