@AspNetSmurfLab 2015-12-08
Here are the numbers from wrk when running the plaintext benchmark with and without this change:

__Baseline:__

	wrk -c 256 -t 32 -d 30 -s ./scripts/pipeline.lua http://10.0.0.100:5001/plaintext -- 16

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    14.00ms   37.09ms 443.35ms   92.45%
	    Req/Sec    21.68k     2.62k  139.04k    89.94%
	  20722668 requests in 30.10s, 2.55GB read
	Requests/sec: 688462.29
	Transfer/sec:     86.67MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    15.18ms   36.59ms 640.33ms   90.76%
	    Req/Sec    21.75k     2.69k  146.35k    90.92%
	  20786049 requests in 30.10s, 2.56GB read
	Requests/sec: 690577.45
	Transfer/sec:     86.93MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency    16.44ms   45.55ms   1.09s    92.51%
	    Req/Sec    21.65k     2.28k  185.05k    88.12%
	  20695354 requests in 30.10s, 2.54GB read
	Requests/sec: 687599.80
	Transfer/sec:     86.56MB


__After merge:__

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     4.26ms    5.24ms 137.84ms   95.14%
	    Req/Sec    21.69k     1.11k   34.26k    73.50%
	  20739007 requests in 30.04s, 2.55GB read
	Requests/sec: 690491.69
	Transfer/sec:     86.92MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     4.24ms    5.42ms 218.06ms   95.36%
	    Req/Sec    21.74k     1.07k   48.29k    72.91%
	  20789200 requests in 30.05s, 2.56GB read
	Requests/sec: 691930.78
	Transfer/sec:     87.10MB

	Running 30s test @ http://10.0.0.100:5001/plaintext
	  32 threads and 256 connections
	  Thread Stats   Avg      Stdev     Max   +/- Stdev
	    Latency     4.21ms    5.03ms 150.64ms   94.96%
	    Req/Sec    21.58k     1.03k   30.87k    69.44%
	  20637557 requests in 30.03s, 2.54GB read
	Requests/sec: 687156.69
	Transfer/sec:     86.50MB

Nothing too significant RPS-wise.