@benaadams 2015-12-08
@halter73 Yeah, but look at latency...!