@benaadams 2015-12-08
I'll take the x4 latency improvement with no detriment to RPS :+1: