@muratg 2015-12-08
@benaadams LogLevel.Verbose was recently renamed to LogLevel.Trace (https://github.com/aspnet/Announcements/issues/124).  Do you have the up-to-date logging packages?