@muratg 2015-12-08
I guess the updated packages were not pushed to the feeds yet for some reason.  cc: @pranavkm