@benaadams 2015-12-08
@muratg timing? Feeds updated 20 hours ago logging change went in 19 hours ago.