@halter73 2015-12-08
We're going to need to get new numbers: https://github.com/aspnet/KestrelHttpServer/pull/458#issuecomment-162992708