@benaadams 2015-12-09
Merged  #427 with this as it contains a regress fix.